(big) Move stuff back into `.cpp` & fix A LOT of things
Sorry I'm lost at this point & I've spent like too many hours here.
I'm glad things finally work.

Circular dependencies, forward declarations & indexing via strings vs
const char pointers were the the most draining.

Let's go yo
